Seed the agent and skill settings forms from the loaded record instead of copying it in through an effect
