+++ packages/twenty-front/src/pages/settings/ai/SettingsAgentForm.tsx
@@ -1,478 +1,77 @@
-import { CombinedGraphQLErrors } from '@apollo/client/errors';
-import { styled } from '@linaria/react';
-import { useParams } from 'react-router-dom';
-import { useDebouncedCallback } from 'use-debounce';
-
-import { SaveAndCancelButtons } from '@/settings/components/SaveAndCancelButtons/SaveAndCancelButtons';
-import { SettingsPageContainer } from '@/settings/components/SettingsPageContainer';
-import { SettingsTabBar } from '@/settings/components/layout/SettingsTabBar';
-import { SettingsRolesQueryEffect } from '@/settings/roles/components/SettingsRolesQueryEffect';
-import { useSaveDraftRoleToDB } from '@/settings/roles/role/hooks/useSaveDraftRoleToDB';
-import { settingsDraftRoleFamilyState } from '@/settings/roles/states/settingsDraftRoleFamilyState';
-import { settingsPersistedRoleFamilyState } from '@/settings/roles/states/settingsPersistedRoleFamilyState';
-import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
-import { SettingsPageLayout } from '@/settings/components/layout/SettingsPageLayout';
-import { activeTabIdComponentState } from '@/ui/layout/tab-list/states/activeTabIdComponentState';
-import { useAtomComponentStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomComponentStateValue';
+import { useQuery } from '@apollo/client/react';
 import { t } from '@lingui/core/macro';
-import { AppPath, SettingsPath } from 'twenty-shared/types';
-import { getSettingsPath, isDefined } from 'twenty-shared/utils';
-import {
-  IconListCheck,
-  IconLock,
-  IconSettings,
-  IconTerminal,
-  useIcons,
-} from 'twenty-ui/icon';
+import { useContext, useEffect } from 'react';
+import { useParams } from 'react-router-dom';
+import { AppPath } from 'twenty-shared/types';
+import { isDefined } from 'twenty-shared/utils';
+import { useIcons } from 'twenty-ui/icon';
 import { Section } from 'twenty-ui/layout';
-import { ThemeContext, themeCssVariables } from 'twenty-ui/theme-constants';
-import { useMutation, useQuery } from '@apollo/client/react';
-import {
-  type CreateAgentInput,
-  CreateOneAgentDocument,
-  FindOneAgentDocument,
-  UpdateOneAgentDocument,
-} from '~/generated-metadata/graphql';
+import { ThemeContext } from 'twenty-ui/theme-constants';
+
+import { SettingsPageContainer } from '@/settings/components/SettingsPageContainer';
+import { SettingsPageLayout } from '@/settings/components/layout/SettingsPageLayout';
+import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
+import { FindOneAgentDocument } from '~/generated-metadata/graphql';
 import { useNavigateApp } from '~/hooks/useNavigateApp';
-import { useNavigateSettings } from '~/hooks/useNavigateSettings';
+import { SettingsAgentDetailSkeletonLoader } from '~/pages/settings/ai/components/SettingsAgentDetailSkeletonLoader';
+import { SettingsAgentFormContent } from '~/pages/settings/ai/components/SettingsAgentFormContent';
+import { getSettingsAiBreadcrumbLinks } from '~/pages/settings/ai/utils/getSettingsAiBreadcrumbLinks';
 
-import { useAtomFamilyStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomFamilyStateValue';
-import { useSetAtomFamilyState } from '@/ui/utilities/state/jotai/hooks/useSetAtomFamilyState';
-import { useContext, useEffect, useState } from 'react';
-import { isDeeplyEqual } from '~/utils/isDeeplyEqual';
-import { SettingsAgentDetailSkeletonLoader } from './components/SettingsAgentDetailSkeletonLoader';
-import { SettingsAgentEvalsTab } from './components/SettingsAgentEvalsTab';
-import { SettingsAgentLogsTab } from './components/SettingsAgentLogsTab';
-import { SettingsAgentRoleTab } from './components/SettingsAgentRoleTab';
-import { SettingsAgentSettingsTab } from './components/SettingsAgentSettingsTab';
-import { SETTINGS_AGENT_DETAIL_TABS } from './constants/SettingsAgentDetailTabs';
-import { useSettingsAgentFormState } from './hooks/useSettingsAgentFormState';
-
-const StyledContentContainer = styled.div`
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  gap: ${themeCssVariables.spacing[8]};
-  width: 100%;
-`;
-
+// Loads the agent and mounts the form only once it is there, so the form
+// starts from the agent's values instead of copying them in afterwards.
 export const SettingsAgentForm = ({ mode }: { mode: 'create' | 'edit' }) => {
   const { theme } = useContext(ThemeContext);
   const { getIcon } = useIcons();
   const { agentId = '' } = useParams<{ agentId: string }>();
-  const navigate = useNavigateSettings();
   const navigateApp = useNavigateApp();
   const { enqueueErrorSnackBar } = useSnackBar();
-  const [isReadonlyMode, setIsReadonlyMode] = useState(false);
-  const [originalFormValues, setOriginalFormValues] = useState<
-    ReturnType<typeof useSettingsAgentFormState>['formValues'] | null
-  >(null);
 
-  const isEditMode = mode === 'edit';
   const isCreateMode = mode === 'create';
 
-  const tabListComponentId = `${SETTINGS_AGENT_DETAIL_TABS.COMPONENT_INSTANCE_ID}-${agentId}`;
-  const activeTabId = useAtomComponentStateValue(
-    activeTabIdComponentState,
-    tabListComponentId,
-  );
-
-  const {
-    formValues,
-    isSubmitting,
-    handleFieldChange,
-    resetForm,
-    setIsSubmitting,
-    validateForm,
-  } = useSettingsAgentFormState(mode);
-
-  const {
-    data,
-    loading,
-    error: agentQueryError,
-  } = useQuery(FindOneAgentDocument, {
+  const { data, loading, error } = useQuery(FindOneAgentDocument, {
     variables: { id: agentId },
     skip: isCreateMode || !agentId,
   });
 
-  useEffect(() => {
-    if (data) {
-      const agent = data?.findOneAgent;
-      if (isDefined(agent)) {
-        if (isDefined(agent.applicationId)) {
-          setIsReadonlyMode(true);
-        }
-        const initialValues = {
-          name: agent.name,
-          label: agent.label,
-          description: agent.description,
-          icon: agent.icon || 'IconLego',
-          modelId: agent.modelId,
-          role: agent.roleId,
-          prompt: agent.prompt,
-          isCustom: agent.isCustom,
-          modelConfiguration: agent.modelConfiguration || {},
-          // TODO: Fallback can be removed once all text response format agents are migrated.
-          responseFormat: agent.responseFormat || { type: 'text' },
-          evaluationInputs: agent.evaluationInputs ?? [],
-        };
-        resetForm(initialValues);
-        setOriginalFormValues(initialValues);
-      } else {
-        enqueueErrorSnackBar({
-          message: t`Agent not found`,
-        });
-        navigateApp(AppPath.NotFound);
-      }
-    }
-  }, [data, resetForm, enqueueErrorSnackBar, navigateApp]);
-
-  useEffect(() => {
-    if (agentQueryError) {
-      enqueueErrorSnackBar({
-        apolloError: agentQueryError,
-      });
-      navigateApp(AppPath.NotFound);
-    }
-  }, [agentQueryError, enqueueErrorSnackBar, navigateApp]);
-
-  const [createAgent] = useMutation(CreateOneAgentDocument);
-  const [updateAgent] = useMutation(UpdateOneAgentDocument);
-
   const agent = data?.findOneAgent;
+  const hasFailedToLoad = !isCreateMode && !loading && !isDefined(agent);
 
-  const settingsDraftRole = useAtomFamilyStateValue(
-    settingsDraftRoleFamilyState,
-    formValues.role || '',
-  );
-  const setSettingsDraftRole = useSetAtomFamilyState(
-    settingsDraftRoleFamilyState,
-    formValues.role || '',
-  );
-  const settingsPersistedRole = useAtomFamilyStateValue(
-    settingsPersistedRoleFamilyState,
-    formValues.role || '',
-  );
-
-  const { saveDraftRoleToDB } = useSaveDraftRoleToDB({
-    roleId: formValues.role || '',
-    isCreateMode: false,
-  });
-
-  const isRoleDirty =
-    isDefined(formValues.role) &&
-    !isDeeplyEqual(settingsDraftRole, settingsPersistedRole);
-
-  const autoSave = useDebouncedCallback(async () => {
-    if (
-      isCreateMode ||
-      isReadonlyMode ||
-      !validateForm() ||
-      isSubmitting ||
-      !agent
-    ) {
+  useEffect(() => {
+    if (!hasFailedToLoad) {
       return;
     }
 
-    const hasChanges =
-      originalFormValues && !isDeeplyEqual(formValues, originalFormValues);
+    enqueueErrorSnackBar(
+      isDefined(error)
+        ? { apolloError: error }
+        : { message: t`Agent not found` },
+    );
+    navigateApp(AppPath.NotFound);
+  }, [hasFailedToLoad, error, enqueueErrorSnackBar, navigateApp]);
 
-    if (!hasChanges && !isRoleDirty) {
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    try {
-      if (isRoleDirty && isDefined(formValues.role)) {
-        try {
-          await saveDraftRoleToDB();
-        } catch (error) {
-          if (CombinedGraphQLErrors.is(error)) {
-            enqueueErrorSnackBar({
-              apolloError: error,
-            });
-          } else {
-            const errorMessage =
-              error instanceof Error ? error.message : String(error);
-            enqueueErrorSnackBar({
-              message: t`Failed to save role permissions: ${errorMessage}`,
-            });
-          }
-          setIsSubmitting(false);
-          return;
-        }
-      }
-
-      await updateAgent({
-        variables: {
-          input: {
-            id: agent.id,
-            name: formValues.name || '',
-            label: formValues.label,
-            description: formValues.description,
-            icon: formValues.icon,
-            modelId: formValues.modelId,
-            roleId: formValues.role,
-            prompt: formValues.prompt,
-            modelConfiguration: formValues.modelConfiguration,
-            responseFormat: formValues.responseFormat,
-            evaluationInputs: formValues.evaluationInputs,
-          },
-        },
-      });
-
-      setOriginalFormValues({ ...formValues });
-    } catch (error) {
-      enqueueErrorSnackBar({
-        apolloError: CombinedGraphQLErrors.is(error) ? error : undefined,
-      });
-    } finally {
-      setIsSubmitting(false);
-    }
-  }, 1_000);
-
-  useEffect(() => {
-    if (isEditMode && !loading && isDefined(originalFormValues)) {
-      autoSave();
-    }
-  }, [
-    formValues,
-    isRoleDirty,
-    isEditMode,
-    loading,
-    originalFormValues,
-    autoSave,
-  ]);
-
-  useEffect(() => {
-    return () => {
-      autoSave.flush();
-    };
-  }, [autoSave]);
-
-  if (!isCreateMode && !loading && !agent) {
-    return null;
+  if (isCreateMode) {
+    return <SettingsAgentFormContent />;
   }
 
-  const canSave = !isReadonlyMode && validateForm() && !isSubmitting;
+  if (!isDefined(agent)) {
+    const AgentIcon = getIcon('IconLego');
 
-  const tabs = [
-    {
-      id: SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.SETTINGS,
-      title: t`Settings`,
-      Icon: IconSettings,
-    },
-    {
-      id: SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.ROLE,
-      title: t`Role`,
-      Icon: IconLock,
-    },
-    {
-      id: SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.EVALS,
-      title: t`Evals`,
-      Icon: IconListCheck,
-    },
-    {
-      id: SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.LOGS,
-      title: t`Logs`,
-      Icon: IconTerminal,
-    },
-  ];
-
-  const handleSave = async () => {
-    if (isReadonlyMode) {
-      return;
-    }
-
-    if (!validateForm()) {
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    try {
-      if (isRoleDirty && isDefined(formValues.role)) {
-        try {
-          await saveDraftRoleToDB();
-        } catch (error) {
-          if (CombinedGraphQLErrors.is(error)) {
-            enqueueErrorSnackBar({
-              apolloError: error,
-            });
-          } else {
-            const errorMessage =
-              error instanceof Error ? error.message : String(error);
-            enqueueErrorSnackBar({
-              message: t`Failed to save role permissions: ${errorMessage}`,
-            });
-          }
-          setIsSubmitting(false);
-          return;
-        }
-      }
-
-      if (isCreateMode) {
-        const input: CreateAgentInput = {
-          name: formValues.name,
-          label: formValues.label,
-          description: formValues.description,
-          icon: formValues.icon,
-          modelId: formValues.modelId,
-          roleId: formValues.role,
-          prompt: formValues.prompt,
-          modelConfiguration: formValues.modelConfiguration,
-          responseFormat: formValues.responseFormat,
-          evaluationInputs: formValues.evaluationInputs,
-        };
-
-        await createAgent({
-          variables: { input },
-        });
-        navigate(SettingsPath.AI);
-        return;
-      }
-
-      if (!agent) {
-        return;
-      }
-
-      await updateAgent({
-        variables: {
-          input: {
-            id: agent.id,
-            name: formValues.name || '',
-            label: formValues.label,
-            description: formValues.description,
-            icon: formValues.icon,
-            modelId: formValues.modelId,
-            roleId: formValues.role,
-            prompt: formValues.prompt,
-            modelConfiguration: formValues.modelConfiguration,
-            responseFormat: formValues.responseFormat,
-            evaluationInputs: formValues.evaluationInputs,
-          },
-        },
-      });
-
-      navigate(SettingsPath.AI);
-    } catch (error) {
-      enqueueErrorSnackBar({
-        apolloError: CombinedGraphQLErrors.is(error) ? error : undefined,
-      });
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const handleCancel = () => {
-    resetForm();
-
-    if (isRoleDirty && isDefined(settingsPersistedRole)) {
-      setSettingsDraftRole(settingsPersistedRole);
-    }
-
-    navigate(SettingsPath.AI);
-  };
-
-  const title = !isCreateMode
-    ? loading
-      ? t`Agent`
-      : agent?.label
-    : t`New Agent`;
-  const AgentIcon = getIcon(formValues.icon || 'IconLego');
-  const breadcrumbText = !isCreateMode
-    ? loading
-      ? t`Agent`
-      : agent?.label
-    : t`New Agent`;
-
-  const isRoleTab = activeTabId === SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.ROLE;
-  const isSettingsTab =
-    activeTabId === SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.SETTINGS;
-  const isEvalsTab = activeTabId === SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.EVALS;
-  const isLogsTab = activeTabId === SETTINGS_AGENT_DETAIL_TABS.TABS_IDS.LOGS;
-
-  const isFormDisabled =
-    isReadonlyMode || (isEditMode ? !agent?.isCustom : false);
-  const isEvalsDisabled =
-    process.env.NODE_ENV === 'development' ? isReadonlyMode : isFormDisabled;
-
-  return (
-    <>
-      <SettingsRolesQueryEffect />
+    return (
       <SettingsPageLayout
-        title={title}
+        title={t`Agent`}
         icon={
           <AgentIcon size={theme.icon.size.md} stroke={theme.icon.stroke.sm} />
         }
-        actionButton={
-          isCreateMode ? (
-            <SaveAndCancelButtons
-              onSave={handleSave}
-              onCancel={handleCancel}
-              isSaveDisabled={!canSave}
-              isLoading={isSubmitting}
-              isCancelDisabled={isSubmitting}
-            />
-          ) : undefined
-        }
-        links={[
-          {
-            children: t`Workspace`,
-            href: getSettingsPath(SettingsPath.General),
-          },
-          { children: t`AI`, href: getSettingsPath(SettingsPath.AI) },
-          { children: breadcrumbText },
-        ]}
-        secondaryBar={
-          isEditMode && loading ? undefined : (
-            <SettingsTabBar
-              tabs={tabs}
-              componentInstanceId={tabListComponentId}
-            />
-          )
-        }
+        links={getSettingsAiBreadcrumbLinks(t`Agent`)}
       >
         <SettingsPageContainer>
           <Section>
-            {isEditMode && loading ? (
-              <SettingsAgentDetailSkeletonLoader />
-            ) : (
-              <StyledContentContainer>
-                {isRoleTab && (
-                  <SettingsAgentRoleTab
-                    formValues={formValues}
-                    onFieldChange={handleFieldChange}
-                    disabled={isFormDisabled}
-                    agentId={agentId}
-                    agentLabel={formValues.label}
-                  />
-                )}
-                {isSettingsTab && (
-                  <SettingsAgentSettingsTab
-                    formValues={formValues}
-                    onFieldChange={handleFieldChange}
-                    disabled={isFormDisabled}
-                    agent={agent}
-                  />
-                )}
-                {isEvalsTab && (
-                  <SettingsAgentEvalsTab
-                    agentId={agentId}
-                    evaluationInputs={formValues.evaluationInputs}
-                    onEvaluationInputsChange={(inputs) =>
-                      handleFieldChange('evaluationInputs', inputs)
-                    }
-                    disabled={isEvalsDisabled}
-                  />
-                )}
-                {isLogsTab && <SettingsAgentLogsTab agentId={agentId} />}
-              </StyledContentContainer>
-            )}
+            <SettingsAgentDetailSkeletonLoader />
           </Section>
         </SettingsPageContainer>
       </SettingsPageLayout>
-    </>
-  );
+    );
+  }
+
+  return <SettingsAgentFormContent key={agent.id} agent={agent} />;
 };
