+++ packages/twenty-front/src/pages/settings/ai/types/SettingsSkillFormValues.ts
@@ -1,0 +1,8 @@
+export type SettingsSkillFormValues = {
+  name: string;
+  label: string;
+  description: string;
+  content: string;
+  icon: string;
+  isLabelSyncedWithName: boolean;
+};
